+++ src/medmain-modal/components/index.js
@@ -1,8 +1,8 @@
-import Confirm from './Confirm';
-import Alert from './Alert';
-import Dialog from './Dialog';
-import DialogButton from './DialogButton';
-import Message from './Message';
-import Modal from './Modal';
+import Confirm from './confirm';
+import Alert from './alert';
+import Dialog from './dialog';
+import DialogButton from './dialog-button';
+import Message from './message';
+import Modal from './modal';
 
 export {Modal, Dialog, DialogButton, Message, Alert, Confirm};
